[arcane,accelerator] Déplace les implémentations de 'ReduceMemoryImpl' et 'RunCommandImpl' dans leur propre fichier.
